added some func

+++ feedChecker/main.py
@@ -1,4 +1,5 @@
 #!/usr/bin/env python
+import datetime
 import feedparser
 import os
 import sys
@@ -18,17 +19,29 @@ def has_argv(check_name):
 def download(uri, file_name):
     urllib.request.urlretrieve(uri, file_name)
 
+def save_log(message):
+    current_dir = os.getcwd()
+    with open(current_dir + '/console.log', mode='a', encoding='utf-8') as fd:
+        now = datetime.datetime.today()
+        msg = str(now) + ' ' + str(message) + '\n'
+        print(msg, end='')
+        fd.write(msg)
+
 def main():
     try:
-        os.remove('feed_old.xml')
-        os.rename('feed.xml', 'feed_old.xml')
+        os.chdir('feedChecker')
+        if( os.path.exists('feed.xml') ):
+            os.remove('feed_old.xml')
+            os.rename('feed.xml', 'feed_old.xml')
         download(uri='https://inside.teu.ac.jp/feed/', file_name='feed.xml')
     except Exception as e:
-        with open('console.log', mode='a', encoding='utf-8') as fd:
-            fd.write(str(e) + '\n')
+        save_log(message=e)
         sys.exit(1)
 
+    fresh_feed = feedparser.parse(r'feed.xml')
     saved_feed = feedparser.parse(r'feed_old.xml')
+    print('[ok] Success to parse feed*.xml')
+
     if( has_argv(check_name="--offline") ):
         print("[info] Offline mode is enabled. Not downlaod feed from remote.")
         feeder = Feeder()
@@ -36,17 +49,18 @@ def main():
         feeder.show(items=['title', 'uri'])
         sys.exit(0)
 
-    fresh_feed = feedparser.parse(r'feed_old.xml')
-    if(fresh_feed == saved_feed):
-        print("[info] Update was not found.")
-        sys.exit(0)
-
     feeder = {'fresh': Feeder(), 'saved': Feeder()}
     feeder['fresh'].set(parsed_feed=fresh_feed)
     feeder['saved'].set(parsed_feed=saved_feed)
+    print('[ok] Success to set parsed_data')
 
     new_entries = Feeder.fetch_diff(target=feeder['fresh'].get(), base=feeder['saved'].get())
     # debug:: print(str(new_entries))
+    if(len(new_entries) < 1):
+         save_log('[info] update was not found.')
+         sys.exit(0)
+    print('[ok] Success to fetch difference between fresh and saved')
+
     for entry in new_entries:
         body = entry['title'] + ' (' + entry['date'] +  ')\n' + entry['uri'] + '\n'
         Feeder.notify_slack(message=body)
